#492 - builder for DataEnvironmentImpl - now in its own class
