+++ src/test/groovy/graphql/DataFetcherTest.groovy
@@ -1,6 +1,5 @@
 package graphql
 
-import graphql.schema.DataFetchingEnvironmentImpl
 import graphql.schema.FieldDataFetcher
 import graphql.schema.GraphQLOutputType
 import graphql.schema.PropertyDataFetcher
@@ -8,6 +7,7 @@ import spock.lang.Specification
 
 import static graphql.Scalars.GraphQLBoolean
 import static graphql.Scalars.GraphQLString
+import static graphql.schema.DataFetchingEnvironmentBuilder.newDataFetchingEnvironment
 
 class DataFetcherTest extends Specification {
 
@@ -55,7 +55,7 @@ class DataFetcherTest extends Specification {
     }
 
     def env(GraphQLOutputType type) {
-        DataFetchingEnvironmentImpl.newDataFetchingEnvironment().source(dataHolder).fieldType(type).build()
+        newDataFetchingEnvironment().source(dataHolder).fieldType(type).build()
     }
 
     def "get field value"() {
