+++ src/test/groovy/graphql/relay/SimpleListConnectionTest.groovy
@@ -1,9 +1,10 @@
 package graphql.relay
 
-import graphql.schema.DataFetchingEnvironmentImpl
 import spock.lang.Specification
 
 import java.nio.charset.StandardCharsets
+
+import static graphql.schema.DataFetchingEnvironmentBuilder.newDataFetchingEnvironment
 
 class SimpleListConnectionTest extends Specification {
     def "invalid list indices handled"() {
@@ -11,7 +12,7 @@ class SimpleListConnectionTest extends Specification {
         def testList = ["a", "b"]
         def listConnection = new SimpleListConnection(testList)
 
-        def env = DataFetchingEnvironmentImpl.newDataFetchingEnvironment().arguments(["after": createCursor(3)]).build()
+        def env = newDataFetchingEnvironment().arguments(["after": createCursor(3)]).build()
 
         when:
         Object item = listConnection.get(env)
@@ -21,7 +22,7 @@ class SimpleListConnectionTest extends Specification {
     }
 
 
-    private String createCursor(int offset) {
+    def createCursor(int offset) {
         def string = SimpleListConnection.DUMMY_CURSOR_PREFIX + Integer.toString(offset)
         return Base64.getEncoder().encodeToString(string.getBytes(StandardCharsets.UTF_8))
     }
